Fix possible integer overflow in WebM duration
The issue could lead to the MPD attribute mediaPresentationDuration being
wrongly generated for WebM streams with a duration longer than INT32_MAX.

The root-cause was that StreamInfo::set_duration() accepted an int instead of a
uint64_t. This seems like a pure typo, since StreamInfo already uses a uint64_t
internally for representing duration.

+++ packager/media/base/stream_info.h
@@ -74,7 +74,7 @@ class StreamInfo : public base::RefCountedThreadSafe<StreamInfo> {
   const std::string& language() const { return language_; }
   bool is_encrypted() const { return is_encrypted_; }
 
-  void set_duration(int duration) { duration_ = duration; }
+  void set_duration(uint64_t duration) { duration_ = duration; }
   void set_codec(Codec codec) { codec_ = codec; }
   void set_codec_config(const std::vector<uint8_t>& data) { codec_config_ = data; }
   void set_codec_string(const std::string& codec_string) {
